refactor: move performanceController into /controllers
